# fix(nmod): Add nmod_ctx to store is_prime