@oscarbenjamin 2024-08-22
> I want to try to get back down to at least the same time if not faster than master.

It is potentially not worth slowing down nmod for prime modulus just to have better error handling for non-prime modulus...