@oscarbenjamin 2024-08-24
> I asked on the Cython mailing list and this is apparently due to cyclic GC:

Using `@cython.no_gc` seems to bring some speedups for some benchmarks.

> Also here are timings for a more macro benchmark where SymPy uses python-flint's `nmod` for `GF(p)` matrices. This is SymPy's pure Python matrix implementation but with `nmod` as the elementary matrix elements. We can invert a 100x100 matrix over `GF(17)` in 11 milliseconds with python-flint master:
> ```python
> In [13]: from sympy import *
>
> In [14]: M = randMatrix(100)
>
> In [15]: dM = M.to_DM(GF(17))
>
> In [26]: %timeit dM.inv()
>11.2 ms ± 164 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
> ```

Turns out I was completely wrong about this. I forgot that in this particular case what happens is that SymPy's sparse implementation converts the matrix to the dense representation which basically means it uses `nmod_mat_inv` to compute the inverse. The fact that is slowed down from 10 milliseconds to 15 milliseconds is still a problem but it is not due to slower arithmetic but rather slower conversions.

This is a better comparison of these using python-flint master:
```python
In [1]: from sympy import *

In [2]: M = randMatrix(100)

In [3]: dM_sympy_sparse = M.to_DM(GF(17)).to_sdm()

In [4]: dM_sympy_dense = M.to_DM(GF(17)).to_ddm()

In [5]: dM_nmod_mat = M.to_DM(GF(17)).to_dfm().rep

In [6]: %timeit dM_sympy_sparse.rref()
136 ms ± 181 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [7]: %timeit dM_sympy_dense.rref()
81.6 ms ± 694 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [8]: %timeit dM_nmod_mat.rref()
313 µs ± 2.19 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```
So in this PR with `no_gc` the pure Python implementation using nmod for the arithmetic is about 300-500x slower than Flint's `nmod_mat_rref` function depending on whether the dense or sparse implementation is used.

If I remove the `no_gc` here:
```diff
diff --git a/src/flint/types/nmod.pyx b/src/flint/types/nmod.pyx
index 3b99b54..145b15d 100644
--- a/src/flint/types/nmod.pyx
+++ b/src/flint/types/nmod.pyx
@@ -195,7 +195,6 @@ cdef class nmod_ctx:
         return self._new(&v)
 
 
-@cython.no_gc
 cdef class nmod(flint_scalar):
     """
     The nmod type represents elements of Z/nZ for word-size n.
```
Then timings are:

With `no_gc`:
```python
In [4]: %timeit dM_sympy_dense.rref()
93.2 ms ± 1.85 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [5]: %timeit dM_sympy_dense.rref()
91.6 ms ± 203 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [6]: %timeit dM_sympy_dense.rref()
94.6 ms ± 2.86 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
Without `no_gc`:
```python
In [8]: %timeit dM_sympy_dense.rref()
110 ms ± 1.95 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [9]: %timeit dM_sympy_dense.rref()
112 ms ± 2.9 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [10]: %timeit dM_sympy_dense.rref()
113 ms ± 4.67 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
With master:
```python
In [4]: %timeit dM_sympy_dense.rref()
80.1 ms ± 280 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [5]: %timeit dM_sympy_dense.rref()
82.1 ms ± 2.3 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [6]: %timeit dM_sympy_dense.rref()
84 ms ± 5.9 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
I've repeated this a few times (different matrices, restart process, etc) and it is consistent so `no_gc` makes a difference of 110 ms vs 90 ms but master is still 80 ms. That means that this PR measures about 10% slower than master but it was 30% slower before adding `@cython.no_gc` so clearly `no_gc` makes a big difference.